requirements.txt
amqp==2.6.1
anyjson==0.3.3
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
dj-database-url==0.5.0
Django==3.1
importlib-metadata==1.7.0
kombu==4.6.11
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
whitenoise==5.2.0
zipp==3.1.0